Remove network-protection-subscription-internal module
