+++ app/build.gradle
@@ -300,7 +300,6 @@ dependencies {
     implementation project(':web-compat-store')
 
     implementation project(':subscriptions-api')
-    implementation project(':network-protection-subscription-internal')
 
     implementation project(':subscriptions-impl')
     internalImplementation project(':subscriptions-internal')
