+++ network-protection/network-protection-impl/src/main/java/com/duckduckgo/networkprotection/impl/subscription/NetPWaitlistEndedChecker.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.networkprotection.subscription
+package com.duckduckgo.networkprotection.impl.subscription
 
 import android.content.SharedPreferences
 import androidx.core.content.edit
