+++ network-protection/network-protection-impl/src/main/res/values/donottranslate.xml
@@ -279,4 +279,17 @@
     <string name="netpDialogVpnAccessRevokedBody">Subscribe to Privacy Pro to reconnect DuckDuckGo VPN.</string>
     <string name="netpDialogVpnAccessRevokedPositiveAction">Subscribe</string>
     <string name="netpDialogVpnAccessRevokedNegativeAction">Dismiss</string>
+
+    <!-- Susbcription -->
+    <string name="netpVerifySubscriptionLabel">Verify Subscription</string>
+    <string name="netpVerifySubscriptionInProgress">Please wait while we verify your subscription…</string>
+    <string name="netpVerifySubscriptionNoSubscription">Oops! You don’t have any valid subscription.</string>
+    <string name="netpVerifySubscriptionNoEntitlement">Oops! Your subscription is not valid for DuckDuckGo VPN.</string>
+
+    <!-- NetP Settings-->
+    <string name="netpSubscriptionSettingsTitle">VPN</string>
+    <string name="netpSubscriptionSettingsConnected">Enabled</string>
+    <string name="netpSubscriptionSettingsConnecting">Connecting…</string>
+    <string name="netpSubscriptionSettingsDisconnected">Disabled</string>
+    <string name="netpSubscriptionSettingsNeverEnabled">Secure your network connection anytime, anywhere</string>
 </resources>
