--- network-protection/network-protection-subscription-internal/.gitignore
@@ -1,1 +1,0 @@
-/build
