--- network-protection/network-protection-subscription-internal/build.gradle
@@ -1,68 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'com.squareup.anvil'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-android {
-    namespace 'com.duckduckgo.networkprotection.subscription'
-    lintOptions {
-        baseline file("lint-baseline.xml")
-    }
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-        }
-    }
-    anvil {
-        generateDaggerFactories = true // default is false
-    }
-}
-
-dependencies {
-    testImplementation 'junit:junit:4.13.1'
-    anvil project(':anvil-compiler')
-    implementation project(':app-build-config-api')
-    implementation project(':anvil-annotations')
-    implementation project(':browser-api')
-    implementation project(':common-ui')
-    implementation project(':common-utils')
-    implementation project(':di')
-    implementation project(':navigation-api')
-    implementation project(':vpn-api')
-    implementation project(':network-protection-api')
-    implementation project(':network-protection-impl')
-    implementation project(':subscriptions-api')
-    implementation project(':settings-api')
-    implementation project(':statistics')
-
-    implementation AndroidX.appCompat
-    implementation AndroidX.lifecycle.runtime.ktx
-    implementation AndroidX.lifecycle.viewModelKtx
-    implementation AndroidX.work.runtimeKtx
-    implementation Google.android.material
-    implementation Google.dagger
-    implementation KotlinX.coroutines.core
-    implementation Square.retrofit2.retrofit
-    implementation Square.retrofit2.converter.moshi
-    implementation "androidx.work:work-multiprocess:_"
-    implementation "com.squareup.logcat:logcat:_"
-
-    // Testing dependencies
-    testImplementation project(':common-test')
-    testImplementation "org.mockito.kotlin:mockito-kotlin:_"
-    testImplementation Testing.junit4
-    testImplementation AndroidX.archCore.testing
-    testImplementation AndroidX.core
-    testImplementation AndroidX.test.ext.junit
-    testImplementation "androidx.test:runner:_"
-    testImplementation Testing.robolectric
-    testImplementation 'app.cash.turbine:turbine:_'
-    testImplementation (KotlinX.coroutines.test) {
-        // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
-        // conflicts with mockito due to direct inclusion of byte buddy
-        exclude group: "org.jetbrains.kotlinx", module: "kotlinx-coroutines-debug"
-    }
-}
