--- network-protection/network-protection-subscription-internal/src/main/res/values/donottranslate.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (c) 2023 DuckDuckGo
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<resources>
-    <!-- Susbcription -->
-    <string name="netpVerifySubscriptionLabel">Verify Subscription</string>
-    <string name="netpVerifySubscriptionInProgress">Please wait while we verify your subscription…</string>
-    <string name="netpVerifySubscriptionNoSubscription">Oops! You don’t have any valid subscription.</string>
-    <string name="netpVerifySubscriptionNoEntitlement">Oops! Your subscription is not valid for DuckDuckGo VPN.</string>
-
-    <!-- NetP Settings-->
-    <string name="netpSubscriptionSettingsTitle">VPN</string>
-    <string name="netpSubscriptionSettingsConnected">Enabled</string>
-    <string name="netpSubscriptionSettingsConnecting">Connecting…</string>
-    <string name="netpSubscriptionSettingsDisconnected">Disabled</string>
-    <string name="netpSubscriptionSettingsNeverEnabled">Secure your network connection anytime, anywhere</string>
-
-</resources>
